refactor: 検索ボックスコンポーネントをhomeページコンポーネントに統合しました。
