+++ src/AnimeList.js
@@ -4,24 +4,20 @@ import {
   View,
   Text,
   TouchableOpacity,
+  TextInput,
   ScrollView,
 } from "react-native";
 import getAnime from "./utils/getAnime";
 import LeftArrow from "./components/LeftArrow";
-import SearchBox from "./components/SearchBox";
 import AnimeCard from "./components/AnimeCard";
 
 const AnimeList = () => {
   const [anime, setAnime] = useState();
   const [isSearch, setIsSearch] = useState(false);
   const [isSort, setIsSort] = useState(false);
+  const [searchText, setSearchText] = useState("");
 
   const subTitle = isSearch ? "結果" : "今人気のアニメ";
-
-  const didMount = () => {
-    getAnime(setAnime);
-  };
-  useEffect(didMount, []);
 
   const backHome = () => {
     setIsSearch(false);
@@ -31,6 +27,25 @@ const AnimeList = () => {
   const openSort = () => {
     setIsSort(true);
   };
+
+  const search = () => {
+    const getAnimeData = {
+      filter: {
+        or: [
+          { property: "title", rich_text: { contains: searchText } },
+          { property: "description", rich_text: { contains: searchText } },
+        ],
+      },
+    };
+
+    setIsSearch(true);
+    getAnime(setAnime, getAnimeData);
+  };
+
+  const didMount = () => {
+    getAnime(setAnime);
+  };
+  useEffect(didMount, []);
 
   return (
     <View style={styles.contianer}>
@@ -43,7 +58,14 @@ const AnimeList = () => {
           <Text style={styles.title}>アニメを検索</Text>
         )}
       </View>
-      <SearchBox setAnime={setAnime} />
+      <TextInput
+        style={styles.searchBox}
+        value={searchText}
+        onChangeText={(text) => setSearchText(text)}
+        onSubmitEditing={search}
+        placeholder="タイトル、またはキーワード"
+        placeholderTextColor={"#808080"}
+      />
       <View style={styles.subTitleContainer}>
         <Text style={styles.subTitle}>{subTitle}</Text>
         <TouchableOpacity style={styles.sortButton} onPress={openSort}>
@@ -75,9 +97,13 @@ const styles = StyleSheet.create({
     fontWeight: "bold",
     color: "white",
   },
-  backHomeButton: {
-    width: 16,
-    aspectRatio: 1,
+  searchBox: {
+    height: 44,
+    paddingHorizontal: 16,
+    borderRadius: 8,
+    fontSize: 16,
+    fontWeight: "600",
+    backgroundColor: "white",
   },
   subTitleContainer: {
     flexDirection: "row",
